whenever calling the function check if it returns "yes"
In, line no. 133 & 156 you had set if wi==True or if wins but the win_check() function returns "yes" or "no", and if wi==True or if wins conditions will run when variable wi and wins are True, but win_check() function returns "yes" or "no", and in Python only 1 = True and 0 = False, other than this whatever you write "yes" or "no" it will not be seen as True because it is in the form of string.
So, I have set if wi=="yes" in line no.133 and if wins=="yes" in line no. 156 now, after a player win's the function will return ( means function will close) and ask the user to play to more game.
Hope, you got the correction. 😃